CSS Reorganization and JS variable scoping.
Organize the CSS file according to shared location, and move magic values in the
JS file to globally scoped variables. Also, add in a background color (this will
certainly change once we can decide on the right color to have). Finally, add
some padding and more structure to the divs displaying the content.