Added a 3rd button to the page.
More work done towards #18. This commit introduces a 3rd tab to the extension
page. It also updates the event handlers and adds a couple of global vars in to
keep track of the state of the page.
Additionally, I've included some example functions for working with line charts
in nvd3.